LT-211, change file structure
