+++ frontend/ui/components/Space/NearProtocol/Network/Accounts/Account/AccountKeys/KeyList/Items/FunctionCall/FunctionCall.module.scss
