+++ frontend/ui/components/Space/NearProtocol/Network/Accounts/Account/AccountKeys/KeyList/Items/FunctionCall/Methods/Methods.module.scss
