+++ frontend/ui/components/Space/NearProtocol/Network/Accounts/Account/AccountKeys/KeyList/Items/Items.jsx
@@ -1,0 +1,43 @@
+import { CopyButton } from '@gc/CopyButton/CopyButton.jsx';
+import { FunctionCall } from './FunctionCall/FunctionCall.jsx';
+import { Tooltip } from '@gc/Tooltip/Tooltip.jsx';
+import { useNavigate, useParams } from 'react-router-dom';
+import cn from './Items.module.scss';
+
+export const Items = ({ value, type }) => {
+  const navigate = useNavigate();
+  const { spaceId, networkId } = useParams();
+
+  const redirectExistKeys = (key) => {
+    if (key.isLocalExists) {
+      navigate(`/space/${spaceId}/near-protocol/${networkId}/keys/${key.publicKey}`);
+    }
+  };
+
+  return (
+    <div
+      className={value.isLocalExists ? cn.itemPointer : cn.item}
+      onClick={() => redirectExistKeys(value)}
+    >
+      <div className={cn.wrapper}>
+        <p className={cn.subtitle}>{value.publicKey}</p>
+        <div className={cn.buttonWrapper}>
+          {value.isLocalExists && (
+            <Tooltip
+              style={cn.tooltip}
+              placement="top"
+              color="black"
+              content="Imported to Lantstool"
+            >
+              <span className={cn.icon} />
+            </Tooltip>
+          )}
+          <div className={cn.copy}>
+            <CopyButton event="onClick" value={value.publicKey} />
+          </div>
+        </div>
+      </div>
+      {type === 'functionCall' && <FunctionCall functionKey={value} />}
+    </div>
+  );
+};
