+++ frontend/ui/components/Space/NearProtocol/Network/Accounts/Account/AccountKeys/KeyList/Items/Items.module.scss
@@ -2,19 +2,6 @@
 @use '@styles/mixins' as *;
 @use '@styles/icons' as *;
 
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.title {
-  @include text-mg-medium;
-  color: $neutral-800;
-  margin: 0;
-  padding: 0 0 8px 0;
-}
-
 @mixin item($pointer) {
   display: flex;
   flex-direction: column;
