+++ frontend/ui/components/Space/NearProtocol/Network/Accounts/Account/AccountKeys/KeyList/KeyList.module.scss
@@ -1,0 +1,15 @@
+@use '@styles/variables' as *;
+@use '@styles/mixins' as *;
+
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.title {
+  @include text-mg-medium;
+  color: $neutral-800;
+  margin: 0;
+  padding: 0 0 8px 0;
+}
